ui: bc: block: bring in getHashDifficulty
